Autocomplete

+++ src/app/pages/filter-page/filter-page.component.ts
@@ -3,6 +3,7 @@ import { FilterService } from '../../core/core-services/filter.service';
 import { Filter } from '../../models/Filter';
 import { allCheckboxFilters } from '../../../data/allCheckboxFilters';
 import { suggestedClients } from '../../../data/suggestedClients';
+import { IOption } from '../../models/IOption';
 
 @Component({
   selector: 'app-filter-page',
@@ -12,9 +13,13 @@ import { suggestedClients } from '../../../data/suggestedClients';
 })
 export class FilterPageComponent implements OnInit {
   private filters: any;
+  private selectedClients: IOption[];
   private checkboxFilters: any = allCheckboxFilters;
   private suggestedClients: any = suggestedClients.map((client) => {
-    return client.name;
+    return {
+      displayName: client.name,
+      value: new Filter({ name: client.name })
+    }
   });
 
   constructor(private filterService: FilterService,
@@ -24,19 +29,35 @@ export class FilterPageComponent implements OnInit {
   ngOnInit() {
     this.filterService.getFilters().subscribe((filters) => {
       this.filters = filters;
+      this.selectedClients = this.mapFiltersToOptions(this.filters.clients);
       this.cd.markForCheck();
     });
   }
 
-  addFilter(category: string, text: string) {
-    const filter = new Filter({
-      name: text
-    });
+  addFilter(category: string, filter: Filter) {
+    if (!(category || !filter)) {
+      return;
+    }
     this.filterService.addFilter(category, filter);
   }
 
   removeFilter(category: string, filter: Filter) {
+    if (!(category || !filter)) {
+      return;
+    }
     this.filterService.removeFilter(category, filter);
+  }
+
+  mapFiltersToOptions(filters: Filter[]): IOption[] {
+    if (!filters) {
+      return;
+    }
+    return filters.map((filter) => {
+      return {
+        displayName: filter.name,
+        value: filter
+      }
+    });
   }
 
   handleCheckboxFilterChange(filterWithCategory: any) {
@@ -46,12 +67,5 @@ export class FilterPageComponent implements OnInit {
     else {
       this.removeFilter(filterWithCategory.category, filterWithCategory.filter)
     }
-  }
-
-  handleClientsFilterAdd(filterWithCategory: any) {
-    if (!filterWithCategory) {
-      return;
-    }
-    this.filterService.addFilter(filterWithCategory.category, filterWithCategory.filter);
   }
 }
